add: confirm delete button yes or no

+++ delete.php
@@ -2,19 +2,40 @@
 
 require_once __DIR__ . ('/utilities/header.php');
 
-if (isset($_GET['id'])) {
+// Vérifie si l'ID du jeu et la confirmation de suppression ont été fournis.
+if (isset($_GET['id']) && isset($_POST['confirm'])) {
 
-// Récupère l'ID du jeu à partir de la superglobale $_GET.
-$currentId = $_GET['id'];
+    // Récupère l'ID du jeu à partir de la superglobale $_GET.
+    $currentId = $_GET['id'];
 
-// Requête SQL pour supprimer le jeu ayant l'ID spécifié.
-$sql = "DELETE FROM jeux_video WHERE ID = $currentId";
+    // Vérifie si l'utilisateur a confirmé la suppression du jeu.
+    if ($_POST['confirm'] === 'Oui') {
 
-// Exécute la requête SQL en utilisant l'objet de connexion à la base de données $dbco
-$deleteRequest = $dbco->query($sql);
+        // Requête SQL pour supprimer le jeu ayant l'ID spécifié.
+        $sql = "DELETE FROM jeux_video WHERE ID = $currentId";
 
-echo "<p>Jeu supprimé avec succès</p>";
+        // Exécute la requête SQL en utilisant l'objet de connexion à la base de données $dbco
+        $deleteRequest = $dbco->query($sql);
 
-echo "<p><a href='/'>Retournez à l'accueil</a></p>";
+        echo "<p>Jeu supprimé avec succès</p>";
+        echo "<p><a href='/'>Retournez à l'accueil</a></p>";
+    } else {
+        // Redirige vers la page d'accueil qui liste les jeux si l'utilisateur clique sur 'Non'.
+        header('Location: /index.php');
+    }
 
+// Si seul l'ID du jeu a été fourni, affiche un formulaire demandant confirmation à l'utilisateur.
+} elseif (isset($_GET['id'])) {
+    
+    // Récupère l'ID du jeu à partir de la superglobale $_GET.
+    $currentId = $_GET['id'];
+    
+    // Affiche le formulaire de confirmation
+    echo "
+        <form method='post'>
+            <p><strong>Voulez vous vraiment supprimer ce jeu ?</strong></p>
+            <input type='submit' name='confirm' value='Oui'>
+            <input type='submit' name='confirm' value='Non'>
+        </form>
+        ";
 }
